Refactor test helpers and add schema test
