+++ packages/postgraphile-core/__tests__/schemas/defaults.test.js
@@ -1,0 +1,3 @@
+import { testSchema } from "../helpers-v5";
+
+it("schema", () => testSchema(__filename, {}));
